Split the structured SCP generator into focused units.
Extract the DAG data structures (nodes, arena, instructions, order) into
sscp_dag.h and the cost function registry (content hashing, bit-packed
verification pool, key assignment) into cost_function_registry.{h,cc}.
Unify the lookup sentinels in one place, rename lookup_sscp_node_cache to
lookup_nodes (it is the authoritative table-to-node mapping, not a cache)
and qualify std calls in headers that only compiled in release builds.
Benchmark neutral.
