+++ src/search/cost_saturation/structured_scp_order_generator.cc
@@ -22,15 +22,6 @@
 using namespace std;
 
 namespace cost_saturation {
-// Mix an (operator, cost) pair into a 64-bit value (splitmix64 finalizer).
-static uint64_t mix_op_cost(int op_id, int cost) {
-    uint64_t x = static_cast<uint64_t>(op_id) * 0x9E3779B97F4A7C15ULL ^
-        static_cast<uint64_t>(static_cast<unsigned int>(cost));
-    x = (x ^ (x >> 30)) * 0xBF58476D1CE4E5B9ULL;
-    x = (x ^ (x >> 27)) * 0x94D049BB133111EBULL;
-    return x ^ (x >> 31);
-}
-
 static const int BITS_PER_WORD = 64;
 
 static int get_num_mask_words(int num_bits) {
@@ -57,7 +48,7 @@ StructuredSCPOrderGenerator::StructuredSCPOrderGenerator(
       log(utils::get_log_for_verbosity(verbosity)) {
     int num_abstractions = this->abstractions.size();
     lookup_tables.resize(num_abstractions);
-    lookup_sscp_node_cache.resize(num_abstractions);
+    lookup_nodes.resize(num_abstractions);
     scf_cache.resize(num_abstractions);
     table_by_restricted_costs.resize(num_abstractions);
     lookup_tables_cache.initialize(num_abstractions);
@@ -244,12 +235,12 @@ StructuredSCPOrder StructuredSCPOrderGenerator::generate() {
     NodeId root_node = create_sscp_order_dag();
     assert(abstractions.size() == unsolvability_infos.size());
     assert(lookup_tables.size() == abstractions.size());
-    assert(lookup_sscp_node_cache.size() == abstractions.size());
+    assert(lookup_nodes.size() == abstractions.size());
     cout << "Recomputed lookup tables: " << recomputed_lookup_tables << endl;
     cout << "Lookup table cache hits: " << lookup_cache_hits << endl;
     cout << "Lookup table cache size: " << lookup_tables_cache.size()
          << endl;
-    cout << "Stored cost functions: " << packed_costs.size() << endl;
+    cout << "Stored cost functions: " << cost_functions.size() << endl;
     int num_lookup_nodes = 0;
     int num_sum_nodes = 0;
     int num_nontrivial_sum_nodes = 0;
@@ -280,11 +271,6 @@ StructuredSCPOrder StructuredSCPOrderGenerator::generate() {
     return create_structured_scp_order(root_node);
 }
 
-// Sentinel table id caching that the lookup node was pruned for these costs.
-static const int PRUNED_LOOKUP = -1;
-// Sentinel for lookup nodes that have not been computed yet.
-static const int UNKNOWN_LOOKUP = -2;
-
 NodeId StructuredSCPOrderGenerator::create_lookup_node(
     const Costs &costs, CostKey cost_key, int abstraction_id) {
     bool cache_this_lookup =
@@ -299,8 +285,8 @@ NodeId StructuredSCPOrderGenerator::create_lookup_node(
             }
             assert(utils::in_bounds(
                        cached_table_id,
-                       lookup_sscp_node_cache[abstraction_id]));
-            return lookup_sscp_node_cache[abstraction_id][cached_table_id];
+                       lookup_nodes[abstraction_id]));
+            return lookup_nodes[abstraction_id][cached_table_id];
         }
     }
     auto cache_table_for_costs = [&](int table_id) {
@@ -330,7 +316,7 @@ NodeId StructuredSCPOrderGenerator::create_lookup_node(
             if (table_id == PRUNED_LOOKUP) {
                 return NO_NODE;
             }
-            return lookup_sscp_node_cache[abstraction_id][table_id];
+            return lookup_nodes[abstraction_id][table_id];
         }
     }
     auto cache_table_for_restricted_costs = [&](int table_id) {
@@ -370,13 +356,13 @@ NodeId StructuredSCPOrderGenerator::create_lookup_node(
                 return NO_NODE;
             }
             cache_table_for_restricted_costs(lookup_table_id);
-            return lookup_sscp_node_cache[abstraction_id][lookup_table_id];
+            return lookup_nodes[abstraction_id][lookup_table_id];
         }
     }
     lookup_tables[abstraction_id].push_back(move(goal_distances));
     NodeId node = nodes.add_lookup_node(abstraction_id, lookup_table_id);
 
-    lookup_sscp_node_cache[abstraction_id].push_back(node);
+    lookup_nodes[abstraction_id].push_back(node);
     if (options.cache_scf_functions) {
         scf_cache[abstraction_id].emplace_back(
             compute_scf(
@@ -396,10 +382,9 @@ NodeId StructuredSCPOrderGenerator::create_lookup_node(
 CostContext StructuredSCPOrderGenerator::make_cost_context(
     Costs &&costs) {
     CostContext context{move(costs), 0, 0, {}, {}};
-    for (size_t op_id = 0; op_id < context.costs.size(); ++op_id) {
-        context.cost_hash += mix_op_cost(op_id, context.costs[op_id]);
-    }
-    context.key = lookup_costs_or_register(context.costs, context.cost_hash);
+    context.cost_hash = CostFunctionRegistry::compute_hash(context.costs);
+    context.key = cost_functions.register_or_lookup(
+        context.costs, context.cost_hash);
     if (options.use_affecting_labels) {
         int num_operators = context.costs.size();
         context.live_ops.assign(get_num_mask_words(num_operators), 0);
@@ -431,6 +416,29 @@ void StructuredSCPOrderGenerator::update_cost_context(
             live_word |= bit;
         }
     }
+}
+
+void StructuredSCPOrderGenerator::reduce_cost_context(
+    CostContext &context, const SaturatedCostFunction &scf) {
+    for (int op_id : scf.nonzero_ops) {
+        int remaining = context.costs[op_id];
+        int saturated = scf.costs[op_id];
+        assert(remaining >= 0);
+        assert(saturated <= remaining);
+        assert(remaining == INF || saturated != INF);
+        // Left addition: x - y = x for all values y if x is infinite.
+        if (remaining != INF) {
+            int reduced = (saturated == -INF) ? INF : remaining - saturated;
+            assert(reduced >= 0);
+            context.costs[op_id] = reduced;
+            context.cost_hash +=
+                CostFunctionRegistry::mix_op_cost(op_id, reduced) -
+                CostFunctionRegistry::mix_op_cost(op_id, remaining);
+        }
+    }
+    update_cost_context(context, scf.nonzero_ops);
+    context.key = cost_functions.register_or_lookup(
+        context.costs, context.cost_hash);
 }
 
 vector<vector<int>>
@@ -550,10 +558,7 @@ void collect_compositional_nodes(
 StructuredSCPOrder StructuredSCPOrderGenerator::create_structured_scp_order(
     NodeId root_node) {
     // Free the construction-time caches before materializing instructions.
-    decltype(cost_key_by_hash)().swap(cost_key_by_hash);
-    utils::release_vector_memory(cost_key_overflow);
-    utils::release_vector_memory(packed_costs.data);
-    utils::release_vector_memory(packed_costs.offsets);
+    cost_functions.release_memory();
     lookup_tables_cache.release_memory();
     utils::release_vector_memory(table_by_restricted_costs);
     utils::release_vector_memory(relevant_op_ids_by_abstraction);
@@ -598,7 +603,7 @@ StructuredSCPOrder StructuredSCPOrderGenerator::create_structured_scp_order(
     int value_id = 0;
     int num_lookup_nodes = 0;
     for (const vector<NodeId> &lookup_nodes_by_abstraction :
-         lookup_sscp_node_cache) {
+         lookup_nodes) {
         for (NodeId node : lookup_nodes_by_abstraction) {
             value_ids[node] = value_id;
             ++value_id;
@@ -685,8 +690,8 @@ void StructuredSCPOrderGenerator::create_compact_lookup_tables() {
             // Remove abstraction without lookup table.
             abstractions.erase(abstractions.begin() + abstr_id);
             lookup_tables.erase(lookup_tables.begin() + abstr_id);
-            lookup_sscp_node_cache.erase(
-                lookup_sscp_node_cache.begin() + abstr_id);
+            lookup_nodes.erase(
+                lookup_nodes.begin() + abstr_id);
             unsolvability_infos.erase(
                 unsolvability_infos.begin() + abstr_id);
             ++removed_abstractions;
@@ -727,130 +732,6 @@ const SaturatedCostFunction &StructuredSCPOrderGenerator::get_saturated_costs(
                 get_lookup_table(node), options.use_general_cp));
         return *scf_scratch;
     }
-}
-
-// Map INF to 0 and finite costs to cost + 1 to avoid a special case.
-static unsigned int packed_cost_value(int cost) {
-    return (cost == INF) ? 0 : static_cast<unsigned int>(cost) + 1;
-}
-
-/* Pack with a compile-time width that divides 8, so the loops have fixed
-   stride and vectorize. */
-template<int BITS>
-static void pack_values(const Costs &costs, vector<uint8_t> &blob) {
-    constexpr int PER_BYTE = 8 / BITS;
-    size_t num_full_bytes = costs.size() / PER_BYTE;
-    size_t offset = blob.size();
-    blob.resize(offset + (costs.size() + PER_BYTE - 1) / PER_BYTE, 0);
-    uint8_t *bytes = blob.data() + offset;
-    for (size_t b = 0; b < num_full_bytes; ++b) {
-        uint8_t value = 0;
-        for (int j = 0; j < PER_BYTE; ++j) {
-            value |= packed_cost_value(costs[b * PER_BYTE + j]) << (j * BITS);
-        }
-        bytes[b] = value;
-    }
-    for (size_t i = num_full_bytes * PER_BYTE; i < costs.size(); ++i) {
-        bytes[num_full_bytes] |=
-            packed_cost_value(costs[i]) << (i % PER_BYTE * BITS);
-    }
-}
-
-template<typename UInt>
-static void pack_values_wide(const Costs &costs, vector<uint8_t> &blob) {
-    size_t offset = blob.size();
-    blob.resize(offset + costs.size() * sizeof(UInt));
-    UInt *values = reinterpret_cast<UInt *>(blob.data() + offset);
-    for (size_t i = 0; i < costs.size(); ++i) {
-        values[i] = static_cast<UInt>(packed_cost_value(costs[i]));
-    }
-}
-
-void PackedCostsPool::pack(const Costs &costs, vector<uint8_t> &blob) {
-    unsigned int max_finite_cost = 0;
-    for (int cost : costs) {
-        assert(cost >= 0);
-        max_finite_cost = max(
-            max_finite_cost,
-            (cost == INF) ? 0 : static_cast<unsigned int>(cost));
-    }
-    uint8_t bits_per_cost = bit_ceil(
-        static_cast<uint8_t>(bit_width(max_finite_cost + 1)));
-
-    blob.clear();
-    blob.reserve(1 + (bits_per_cost * costs.size() + 7) / 8);
-    blob.push_back(bits_per_cost);
-    switch (bits_per_cost) {
-    case 1:
-        pack_values<1>(costs, blob);
-        break;
-    case 2:
-        pack_values<2>(costs, blob);
-        break;
-    case 4:
-        pack_values<4>(costs, blob);
-        break;
-    case 8:
-        pack_values_wide<uint8_t>(costs, blob);
-        break;
-    case 16:
-        pack_values_wide<uint16_t>(costs, blob);
-        break;
-    default:
-        pack_values_wide<uint32_t>(costs, blob);
-        break;
-    }
-}
-
-void StructuredSCPOrderGenerator::reduce_cost_context(
-    CostContext &context, const SaturatedCostFunction &scf) {
-    for (int op_id : scf.nonzero_ops) {
-        int remaining = context.costs[op_id];
-        int saturated = scf.costs[op_id];
-        assert(remaining >= 0);
-        assert(saturated <= remaining);
-        assert(remaining == INF || saturated != INF);
-        // Left addition: x - y = x for all values y if x is infinite.
-        if (remaining != INF) {
-            int reduced = (saturated == -INF) ? INF : remaining - saturated;
-            assert(reduced >= 0);
-            context.costs[op_id] = reduced;
-            context.cost_hash += mix_op_cost(op_id, reduced) -
-                mix_op_cost(op_id, remaining);
-        }
-    }
-    update_cost_context(context, scf.nonzero_ops);
-    context.key = lookup_costs_or_register(context.costs, context.cost_hash);
-}
-
-/* Return the key under which the given cost function is registered, so
-   that all other hash maps can use the small key instead of the full cost
-   function. */
-CostKey StructuredSCPOrderGenerator::lookup_costs_or_register(
-    const Costs &costs, uint64_t hash) {
-    auto it = cost_key_by_hash.find(hash);
-    if (it == cost_key_by_hash.end()) {
-        CostKey key = packed_costs.size();
-        PackedCostsPool::pack(costs, packed_costs_scratch);
-        packed_costs.append(packed_costs_scratch);
-        cost_key_by_hash.emplace(hash, key);
-        return key;
-    }
-    PackedCostsPool::pack(costs, packed_costs_scratch);
-    if (packed_costs.equals(it->second, packed_costs_scratch)) {
-        return it->second;
-    }
-    // Hash collision: look for the cost function in the overflow list.
-    for (const auto &[overflow_hash, overflow_key] : cost_key_overflow) {
-        if (overflow_hash == hash &&
-            packed_costs.equals(overflow_key, packed_costs_scratch)) {
-            return overflow_key;
-        }
-    }
-    CostKey key = packed_costs.size();
-    packed_costs.append(packed_costs_scratch);
-    cost_key_overflow.emplace_back(hash, key);
-    return key;
 }
 
 void add_structured_order_generator_options_to_parser(
